MemMappedBucket: Ensure memory is zeroed on bucket creation.
Since it's possible for memory to be clobbered past the finalised end of
the previous bucket, we ensure that the address for the first element is
zeroed out.

Obtaining the next element has also been simplified to avoid walking the
complete bucket on each request for the next element to populate.